Adjust tcp_listen to just return the sock_fd.
Filling out a Connection will be useful, but at this point we don't have
enough information to fill it in. We don't get the client ip+port until
we return from accept(), and we actually don't get the server ip until
then either because we're listening on a wildcard IP addr.